separate merge tests
